Crypto

Description

This symmetric cryptosystem works as follows:

AES-256 is used as a specification of choice.

Initialisation vector (IV) is generated randomly for each encryption.

AES key is made by taking a SHA-256 hash of the provided master password.

A checkfile is created upon initialising a password storage directory. This checkfile contains 16 bytes of pre-defined noise (dummy values) which are encrypted using our AES key and a newly generated IV.

Checking if the master password is correct consists of

decrypting the checkfile with master password's AES key

comparing the result with dummy values.

The check is passed if the comparison yield equalness of both values.

Working with checkfiles

checkMasterPass :: FilePath -> ByteString -> IO Bool

Checks if the master password entered by user is correct.

checkFileExists :: FilePath -> MaybeT IO FilePath

Tries to find the "nearest" checkfile by sequentially going up a directory until hass core directory is reached.

Returns a wrapped Nothing if no checkfile was found, or a wrapped Just FilePath in case of a success.

createStorage :: FilePath -> IO ()

Check if the directory is already initialised as a storage and, if not, create a checkfile.

createCheckFile :: FilePath -> ByteString -> IO ()

Create a checkfile in the specified directory using provided master password.

onEmptyMP :: IO Bool

Prompt for confirmation on entering empty master password and parse (parseConfirmationRequest) it. Defaulting to negative.

Encryption functions

reencrypt :: TraverseReqs -> ByteString -> ByteString

Reencrypt a ByteString by decrypting it with old and encrypting with new data.

encrypt :: CryptReqs -> ByteString -> ByteString

Encrypt a ByteString.

decrypt :: CryptReqs -> ByteString -> ByteString

Decrypt a ByteString.