Daemon

Description

A simple dedicated daemon used to efficiently manage password-to-clipboard selections. After start it waits a number of seconds (45 by default) and simply sets the clipboard to an empty string.

Documentation

setClipboardAndDaemonize :: String -> IO ()

Copy a string to the clipboard, starting a daemon on the way to ensure the selection gets erased after a certain delay.

daemonAliveAtClip :: IO ()

To be used before setting the clipboard when hass is invoked with --clip.

startDaemon :: IO ()

Start our daemon.