Git

Description

Helper functions to be used with hass git command.

Currently git repository is initialised in core directory and it follows that only the whole store can be tracked.

Documentation

gitDir :: IO String

Return path to .git directory in hass storage directory.

gitInit :: [String] -> IO ()

If git init is run, add all password files to the repository in an initial commit.

gitAdd :: FilePath -> IO ()

Add a file/directory's contents to the repository and tree.

gitRm :: FilePath -> IO ()

Remove a file/directory's contents from the repository and tree.

gitCommit :: String -> IO ()

Make a commit to the password-storage repository.

gitTrackNewPassword :: FilePath -> String -> IO ()

Combines git-add and git-commit to simplify tracking new or edited password files.

isGitRepo :: IO Bool

Is password storage directory a git repository?

callGit :: [String] -> IO ()

Call git process with passed arguments