Helpers

Description

Helper functions to be used in the main module.

Documentation

runPwgen :: Integer -> Bool -> IO ByteString

Call pwgen and read the resulting generated password. Correctly passes --no-symbols flag.

Traverse functions

Functions to be used with traverseStore

reencryptEntry :: TraverseFunction

Reencrypt entry. To be used with traverseStore function when hass is invoked with init --reencrypt.

copyEntry :: TraverseFunction

Copy file of directory to the location specified in TraverseReqs. To be used with traverseStore function when hass is invoked with cp or mv.

removeEntry :: TraverseFunction

Remove file or directory. To be used with traverseStore function when hass is invoked with rm.

printEntryName :: TraverseFunction

Pretty print entry name. To be used with traverseStore function when hass is invoked with list.

printWith :: TraverseFunction

Apply a filter passed as string within TraverseReqs to the entry name before printing it. To be used with traverseStore function when hass is invoked with find.

grepInPass :: TraverseFunction

Decrypt a password, grep inside it for search string, and display the line containing matched string along with filename. To be used with traverseStore function when hass is invoked with grep.

Main actions on passwords

actionInsert :: FilePath -> CryptReqs -> IO ()

Asks for a password to be inserted, asking to reconfirm unless --echo OR --multiline has been passed, and writes it to the corresponding file. If hass was invoked with --multiline flag, waits for Ctrl+D to stop reading the input.

actionGenerate :: FilePath -> Integer -> CryptReqs -> IO ()

Generates a password of given length using runPwgen. The function then writes the encrypted result to the corresponding file (possibly only replacing the first line if --in-place was specified) and either outputs it to stdout stream, or copies it to the clipboard if --clip was passed.

actionEdit :: FilePath -> CryptReqs -> IO ()

Calls the editor of choice of the system, looking up the $EDITOR environment variable. Falls back to vi as it should be pre-installed on every Linux machine. Calls shred on the temporary file when the editing is done.

actionShow :: FilePath -> CryptReqs -> IO ()

Either outputs the decrypted password to stdout stream, or copies it to the clipboard if --clip was passed.

actionRemove :: FilePath -> IO ()

Removes the password file/directory after asking for confirmation defaulting to negative unless --force was specified. Does not remove non-empty directories unless invoked with --recursive flag. Directory is considered empty if it contains nothing but possibly a .check file.

actionCopy :: FilePath -> FilePath -> Bool -> CryptReqs -> IO ()

Copies the password file/directory to the specified folder. Confirmation for overwriting existing targets defaults to negative and occurs unless --force was specified. After copy execution selectively reencrypts copied files after asking for the most suitable master password.

actionGit :: [String] -> IO ()

Calls git with arguments specified in the command line upon invokation.