Options

Description

This module contains functionality to parse command line arguments with which hass was invoked. All options that can't be needed by the first command are stripped.

Documentation

commands :: IO [String]

Extract commands from command line.

args :: IO [Option]

Extract flags from command line.

inArgs :: Option -> IO Bool

Checks if a flag was passed to hass via command line.

putUsage :: IO ()

Print usage to stderr stream.

usage :: String

A string describing commands available and pattern for invoking hass.

commandUsage :: String -> String

Returns a string showing usage pattern and available flags for a particular command.