Types

Description

Types module contains all self-defined data types and instances, default representations for records, some type synonyms and lenses. This makes this module necessary to be imported by almost all other hass modules.

Documentation

data Option

Data type for parsing command line flags.

Constructors

Reencrypt

Path

Clip

Echo

NoSymbols

InPlace

MultiLine

Force

Recursive

Help

data AoD

Data type for ternary choices - action on directories in traverseStore in this case.

Constructors

No

Before

After

before :: AoD -> Bool

Used to check if aod is Before.

after :: AoD -> Bool

Used to check if aod is After.

data ActionReqs

Record used for efficient passing of arguments to actionOnNonExistingPasses and actionOnExistingPasses functions.

Constructors

ActionReqs

Fields

_entity :: String

Which action to execute?

_dest :: FilePath

Destination file path needed by mv and cp

_passlength :: Integer

Password length needed by actionGenerate

_cr :: CryptReqs

Attributes needed by encrypt-decrypt logic

defAR :: ActionReqs

Default configuration of ActionReqs.

data TraverseReqs

Record used for efficient passing of arguments to traverseStore function. If onCrypt is specified, traverseStore will skip the subfolders which have their own checkfiles.

Constructors

TraverseReqs

Fields

_aod :: AoD

Execute the action on directory entries?

_aocf :: Bool

Execute the action on checkfile entries?

_onCrypt :: Bool

Does the action involve crypto shenanigans?

_cdir :: FilePath

Current directory to pass to actions (will virtually change during traverseStore run)

_target :: FilePath

Target directory to pass to actions (will virtually change during traverseStore run)

_depth :: Integer

Depth of recursion

_act :: TraverseReqs -> FilePath -> IO ()

Action to be executed

_string :: String

An additional string, can be used e.g. in grep or find actions

_oldCR :: CryptReqs

Old master password

_newCR :: CryptReqs

New master password

defTR :: TraverseReqs

Default configuration of TraverseReqs.

data CryptReqs

Record used for efficient passing of arguments to encrypt-decrypt functions.

Constructors

CryptReqs

Fields

_mp :: ByteString

_iv :: ByteString

defCR :: CryptReqs

Default implementation of CryptReqs.

type TraverseFunction = TraverseReqs -> FilePath -> IO ()

Type synonym for better readability.

newtype MaybeT m a

A simple monad transformer.

Constructors

MaybeT

Fields

runMaybeT :: m (Maybe a)

Labels for lenses

TraverseReqs, ActionReqs, CryptReqs

mp :: forall cat. ArrowApply cat => Lens cat CryptReqs ByteString

iv :: forall cat. ArrowApply cat => Lens cat CryptReqs ByteString

passlength :: forall cat. ArrowApply cat => Lens cat ActionReqs Integer

entity :: forall cat. ArrowApply cat => Lens cat ActionReqs String

dest :: forall cat. ArrowApply cat => Lens cat ActionReqs FilePath

cr :: forall cat. ArrowApply cat => Lens cat ActionReqs CryptReqs

target :: forall cat. ArrowApply cat => Lens cat TraverseReqs FilePath

string :: forall cat. ArrowApply cat => Lens cat TraverseReqs String

onCrypt :: forall cat. ArrowApply cat => Lens cat TraverseReqs Bool

oldCR :: forall cat. ArrowApply cat => Lens cat TraverseReqs CryptReqs

newCR :: forall cat. ArrowApply cat => Lens cat TraverseReqs CryptReqs

depth :: forall cat. ArrowApply cat => Lens cat TraverseReqs Integer

cdir :: forall cat. ArrowApply cat => Lens cat TraverseReqs FilePath

aod :: forall cat. ArrowApply cat => Lens cat TraverseReqs AoD

aocf :: forall cat. ArrowApply cat => Lens cat TraverseReqs Bool

act :: forall cat. ArrowApply cat => Lens cat TraverseReqs (TraverseReqs -> FilePath -> IO ())