UtilFunctions

Description

Utility functions to be used in other hass modules.

Documentation

die :: String -> IO ()

Output an error message to stderr stream and exit with exit code 1. Should be used only at user-interaction level (such as prompting for master password), not instead of exceptions.

getByteLine :: IO ByteString

The equivalent of getLine for ByteStrings.

toClipboard :: ByteString -> IO ()

Sets clipboard, starting a daemon to wait 45 seconds and then erase the clipboard.

hassDir :: IO FilePath

The hass directory containing password storages.

passPath :: String -> IO FilePath

Exports path to concrete password file or directory.

makeRelativeToHass :: FilePath -> IO FilePath

Extract only the part of the path relative to hass password storage directory. Does not resolve special entries like ...

checkSneakyPaths :: FilePath -> Bool

Checks if a sneaky path was entered, such as containing ...

listDirectory :: FilePath -> IO [FilePath]

List directory entries, skipping dot files. To be used in listing action.

listRemovingDirectory :: FilePath -> IO [FilePath]

List all directory entries except ., .. and .check. To be used during removal to correctly check if the directory is empty of needed files.

createStorageDir :: String -> IO ()

Creates path to storage directory, creating missing parents directories on demand.

createPath :: FilePath -> IO ()

Creates directories in a path to a file, extracting everything prior to the last slash.

traverseStore :: TraverseReqs -> FilePath -> IO ()

Traverse a directory, recursively executing the passed action on each entry. Executes the action on each directory entry, too, if anything other that No is passed as aod field in TraverseReqs.

prompt :: FilePath -> IO Bool

Prompt for confirmation in the event of trying to overwriting a target.

parseConfirmationRequest :: Bool -> String -> Bool

Parse confirmation request. Takes a default value and a string to be parsed. "y", "yes", "Y" are recognised as confirmative answers, "n", "no", "N" as negative, empty line and unrecognised options will be defaulted.

Actions with temporary effects

withEcho :: Bool -> IO a -> IO a

Set echoing stdin, execute an action and then restore the echo setting to the old value. Restores the old value in case of an exception during action's execution.

withTempFile :: (FilePath -> Handle -> IO a) -> IO a

Execute an action using temporary file. Tries writing to /dev/shm in order to avoid writing to difficult-to-erase disk sectors. Falls back to default $TMPDIR or /tmp if that variable isn't set and asks for confirmation.

withDirectory :: FilePath -> IO a -> IO a

Change the working directory and restore it afterwards, even if the given action fails due to an exception.

If-functions

(?) :: Bool -> a -> a -> a infixr 1

If-function to avoid using the built-in if-then-else construct.

ifOn :: a -> a -> Bool -> a

If-function with boolean argument passed last. To be used in monad bindings.

ifM :: Monad m => m Bool -> m a -> m a -> m a

If-function operating on monadic values.

whenM :: Monad m => m Bool -> m () -> m ()

when-function operating on monadic values.

(?..) :: Maybe a -> b -> (a -> b) -> b infixr 1

The same as maybe, but with "on Just" option passed last as to increase readability.